Export GET-DEFAULT-TEMPORARY-DIRECTORY
I want to run unit tests which create directory and file structures on hard drive. Because of this, I would like to make a temporary directory instead of a temporary file.

I see that CL-FAD already has a function for returning such a directory, but its symbol is not external.